Using proper int types

+++ include/core/control_message.h
@@ -1,12 +1,14 @@
 #pragma once
 
+#include <stdint.h>
+
 #include "core/loop_track.h"
 
 typedef struct ControlMessage {
-  int track_number;
+  uint8_t track_number;
   LoopTrackAction action;
 } ControlMessage;
 
-ControlMessage *cm_create(int track_number, LoopTrackAction action);
+ControlMessage *cm_create(uint8_t track_number, LoopTrackAction action);
 
 void cm_destroy(ControlMessage *cm);
