+++ src/core/control_message.c
@@ -1,4 +1,5 @@
 #include <stdlib.h>
+#include <stdint.h>
 #include <assert.h>
 
 #include "dbg.h"
@@ -6,7 +7,7 @@
 #include "core/control_message.h"
 #include "core/loop_track.h"
 
-ControlMessage *cm_create(int track_number, LoopTrackAction action) {
+ControlMessage *cm_create(uint8_t track_number, LoopTrackAction action) {
   ControlMessage *cm = malloc(sizeof(ControlMessage));
   check_mem(cm);
   cm->action = action;
